Fix paddle centering
From some reason pygame.event.set_grab(True) breaks centering of the
paddle. Once this is set to False it works fine without any issues.
Unfortunately now it is possible to go outside of game screen with mouse
which glitches the paddle. It is not that problematic but still
annoying.